requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
Pillow
psycopg2-binary
django-filter